Completed model consolidation for State, Controller, County, Property, and USMap
