+++ server/src/routes/main.routes.ts
@@ -2,8 +2,8 @@ import { Router } from 'express';
 import { auth } from '../middleware/auth';
 import { Property } from '../models/property.model';
 import { State } from '../models/state.model';
-import { County } from '../models/County';
-import { USMap } from '../models/USMap';
+import { County } from '../models/county.model';
+import { USMap } from '../models/usmap.model';
 import express from 'express';
 
 const router = Router();
