+++ server/src/utils/controller-utils.ts
@@ -1,7 +1,7 @@
 import { ObjectId } from 'mongodb';
 import mongoose from 'mongoose';
 import { Controller } from '../models/controller.model';
-import { County } from '../models/County';
+import { County } from '../models/county.model';
 import logger from '../utils/logger';
 
 /**
